[Refactoring] Move FillProtocolStubs to its own file
